got rid of the hardcoding part of broker

+++ docker/cav/vehicle/run.sh
@@ -1,5 +1,12 @@
 #! /bin/bash
 
+# check if broker IP file exists
+if [ ! -e 'bip/ip.txt' ]
+then
+	echo 'bip/ip.txt missing'
+	exit 1
+fi
+
 # delete all stopped containers
 docker rm $(docker ps -a -q)
 
@@ -18,7 +25,12 @@ mkdir feedback
 
 # run docker containers
 cd shared_folder
-docker run -v `pwd`:/mydata -p 9000:9000 -p 3000:3000 -d vehicle/expose /mydata 
+#docker run -v `pwd`:/mydata -p 9000:9000 -p 3000:3000 -d thatape/ecf_vehicle:expose /mydata 
+docker run -v `pwd`:/mydata -p 9000:9000 -p 3000:3000 -d st571332/ecf_vehicle:expose /mydata 
+
+cd ../bip
+brokerip_path=`pwd`
 
 cd ../feedback
-docker run --network host -v `pwd`:/output -d vehicle/feedback
+#docker run --network host -v `pwd`:/output -v $brokerip_path:/bip -d thatape/ecf_vehicle:feedback
+docker run --network host -v `pwd`:/output -v $brokerip_path:/bip -d st571332/ecf_vehicle:feedback
